% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{report_period}
\alias{report_period}
\title{report_period}
\usage{
report_period(from_timestamp, to_timestamp)
}
\arguments{
\item{from_timetamp, to_timestamp}{timestamps for the begin and end of the report}
}
\description{
Print a formatted from-to sentence
}
