% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{wrangle_devices}
\alias{wrangle_devices}
\title{wrangle_devices}
\usage{
wrangle_devices(filtered_devices)
}
\arguments{
\item{filtered_devices}{a data.frame or path to file}
}
\description{
wrangle the devices data
}
\examples{
wrangle_devices(test_params()$filtered_devices)
}
