% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.R
\name{make_nz_datetime}
\alias{make_nz_datetime}
\title{make_nz_datetime}
\usage{
make_nz_datetime(timestamp)
}
\arguments{
\item{timestamp}{a Unix Timestamp}
}
\description{
convert a time stamp into a NZ timezone datetime
}
\examples{
make_nz_datetime(1555555555)
}
